Cosmetic cleanup & support for older gettext.

+++ src/stdlib/mystring.c
@@ -681,7 +681,7 @@ char *
 llstrapp (char *dest, size_t limit, const char *src)
 {
 	size_t len = strlen(dest);
-	size_t n = limit-len;
+	int n = limit-len;
 	ASSERT(n >= 0);
 	if (n < 2) /* must fit trailing zero */
 		return dest;
